+++ js/data.js
@@ -1,0 +1,51 @@
+const TITLES = [
+  'Скромная квартира',
+  'Роскошная квартира',
+  'Апартаменты',
+  'Хибара',
+  'Сарай',
+  'Квартира',
+  'Дом',
+  'Гараж',
+  'Ангар',
+  'Хостел',
+  'Стойло',
+];
+
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const DESCRIPTIONS = [
+  'Просторная квартира со стульями',
+  'Уютные комнаты',
+  'Светлая и чистая',
+  'Просторный балкон',
+  'Кладовка включена',
+  'Потолки высокие',
+  'Фото-обои на стенах',
+  'Всюду ковры и чучела',
+  'Большой телевизор и своя стиральная машина',
+  'До метро 5 минут',
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+export {TITLES, TYPES, FEATURES, DESCRIPTIONS, PHOTOS};
